+++ src/main/java/roomescape/controller/dto/reservationtime/AvailableReservationTimesQuery.java
@@ -1,4 +1,4 @@
-package roomescape.controller.dto;
+package roomescape.controller.dto.reservationtime;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeParseException;
