+++ src/main/java/roomescape/controller/dto/theme/ThemeRankResponse.java
@@ -1,4 +1,4 @@
-package roomescape.controller.dto;
+package roomescape.controller.dto.theme;
 
 public record ThemeRankResponse(
         int rank,
