+++ src/main/java/roomescape/service/ReservationService.java
@@ -6,8 +6,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import roomescape.domain.ReservedTimes;
 import roomescape.service.dto.CreateReservationCommand;
-import roomescape.controller.dto.ReservationPagingQuery;
-import roomescape.controller.dto.ReservationResponse;
+import roomescape.controller.dto.reservation.ReservationPagingQuery;
+import roomescape.controller.dto.reservation.ReservationResponse;
 import roomescape.domain.Reservation;
 import roomescape.domain.ReservationTime;
 import roomescape.domain.Theme;
