+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -4,11 +4,11 @@ import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import roomescape.controller.dto.AvailableReservationTimeResponse;
-import roomescape.controller.dto.AvailableReservationTimesQuery;
-import roomescape.controller.dto.AvailableReservationTimesResponse;
-import roomescape.controller.dto.ReservationTimeResponse;
-import roomescape.controller.dto.ThemeResponse;
+import roomescape.controller.dto.reservationtime.AvailableReservationTimeResponse;
+import roomescape.controller.dto.reservationtime.AvailableReservationTimesQuery;
+import roomescape.controller.dto.reservationtime.AvailableReservationTimesResponse;
+import roomescape.controller.dto.reservationtime.ReservationTimeResponse;
+import roomescape.controller.dto.theme.ThemeResponse;
 import roomescape.domain.ReservationTime;
 import roomescape.domain.ReservedTimes;
 import roomescape.global.exception.DuplicateReservationTimeException;
