+++ src/main/java/roomescape/service/ThemeService.java
@@ -6,10 +6,9 @@ import java.util.stream.IntStream;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import roomescape.controller.dto.ThemeRequest;
-import roomescape.controller.dto.ThemeRankResponse;
-import roomescape.controller.dto.ThemeRankingQuery;
-import roomescape.controller.dto.ThemeResponse;
+import roomescape.controller.dto.theme.ThemeRankResponse;
+import roomescape.controller.dto.theme.ThemeRankingQuery;
+import roomescape.controller.dto.theme.ThemeResponse;
 import roomescape.domain.Theme;
 import roomescape.repository.ThemeRepository;
 import roomescape.service.dto.CreateThemeCommand;
